feat: userType을 user 도메인으로 이동시키고 불필요한 코드 파일 삭제
