+++ src/frontend/src/api/user/user.shcema.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod';
+
+export const UserSchema = z.object({
+  id: z.number().int(),
+  email: z.string().email(),
+  name: z.string(),
+});
+
+export type UserType = z.infer<typeof UserSchema>;
