--- src/frontend/src/types/user.d.ts
@@ -1,7 +1,0 @@
-export type UserType = {
-  id: number;
-  email: string;
-  name: string;
-  createdAt: string;
-  modifiedAt: string;
-};
